duits/requirements.txt
Django==4.2
django-crispy-forms
bootstrap4
